Generalize commit:f707c6b, assuming no velocity/acceleration in current state
In addition to takeoff, let the goto and waypoint activities assume that the
current state is hovering (no velocity). Otherwise, the trajectory generation
computes weird trajectories depending on the residual noise present in the
state estimation.

However, note that for a waypoint concatenated to a previous trajectory not yet
executed, the complete start state is used (since it comes from a previous
planning, is is noise free).