Add a replay() function, playing back a log file (or any trajectory file)
The input file must have the same format as log files, that is:
 " ts x y z yaw vx vy vz wz ax ay az "

The trajectory can be sampled at any frequency. If the delta between two
samples in the input file is greater than the exec task frequency, the local
planner is called to interpolate between the samples. On the other hand, if the
delta is too small, oversampled entries are simply discarded.