Fix the semantic of the interrupt codel in the waypoint activity
The waypoint activity is not meant to be interrupted by itself, since it's
meant to append paths. However, the "interrupt" keyword is used here to wait
that a previous instance has completed its computations before the next one
starts (so that the "current" state is consistent).

However, the "stop" codel was missing, so the activity was just leaving without
updating the trajectory. Just point the stop codel to the codel pushing the
trajectory, this will DTRT.

Fixes issue #147