Add a velocity planner and a "stop" service planning a minimum time 0 speed
Introduce a second planner used for velocity planning. The position() of
the planner is interpreted as a velocity, so that a minimum time trajectory
is generated to reach a desired velocity. The resulting position is then
integrated numerically.

A new 'stop' activity uses this to plan a minimum time stop from the current
reference.

Also extend the 'exec' task so that it is able to integrate a current
reference with non zero velocity. This can happen if one requests a
'waypoint' to be reached with a non zero velocity (or acceleration).
The trajectory data in the IDS had to be extended to hold the current vel, acc
and jerk from the planner so that the integration for "velocity" trajectories
work as planned.

The log file is also extended to report acceleration and jerk, since they are
readily available for logging.