Rearrange exec task state machine to guarantee schedulability
As it was written, if was theoretically possible for the exec task to loop
in the <wait> state without ever pausing, thus overshooting the period.
In practice, the codels were written so that this could not happen, but better
fix that in order to makes things cleaner.

It is enough to add a new <over> state that accomplishes what was done in the
<wait> state when it was scheduled without a pause, i.e. at the end of a
planned or servoed motion. The main purpose of this is to create a non-paused
transition after the execution so that the plan task is awaken when it waits
for that event.

Spotted by Felix & Tina after a tough trace analysis session!
No human were harmed though.